Refactor color palette functionality into separate methods
