--- crimsobot/data/img/colors.json
@@ -1,31 +1,0 @@
-{
-    "#000000": ":black_large_square:",
-    "#474757": ":new_moon_with_face:",
-    "#6c6c6b": ":elephant:",
-    "#b5bab1": ":clock4:",
-    "#f4fee7": ":white_large_square:",
-    "#5e322a": ":briefcase:",
-    "#955219": ":chestnut:",
-    "#df6621": ":tangerine:",
-    "#f1ce17": ":lemon:",
-    "#bc8b57": ":cookie:",
-    "#ebcd93": ":moneybag:",
-    "#0e5821": ":deciduous_tree:",
-    "#086f93": ":melon:",
-    "#00a3c3": ":droplet:",
-    "#44ae2a": ":green_heart:",
-    "#9de133": ":white_check_mark:",
-    "#493149": ":eggplant:",
-    "#f8bf87": ":moneybag:",
-    "#d22c58": ":strawberry:",
-    "#0000ff": ":sweat_drops:",
-    "#850000": ":crab:",
-    "#ffcc4d": ":kissing_smiling_eyes:",
-    "#71a3c8": ":bullettrain_front:",
-    "#9266cc": ":octopus:",
-    "#000060": ":flag_eu:",
-    "#f8afff": ":pig2:",
-    "#ff0000": ":tomato:",
-    "#00ffff": ":dress:",
-    "#b2ffba": ":pear:"
-}
